fix codecheck & bug

+++ attention/incre_flash_attention/op_kernel/arch32/incre_flash_attention_preload_mla.h
@@ -1991,8 +1991,8 @@ __aicore__ inline void IncreFlashAttentionAttenPreloadMla<IFAT>::AttenMaskCopyFo
     } else {
         headGCount = dealRowCount;
     }
-    uint32_t dstMaskOffset = 0;
-    uint32_t srcMaskBaseOffset = 0;
+    uint64_t dstMaskOffset = 0;
+    uint64_t srcMaskBaseOffset = 0;
     // head
     SetMaskCount();
     SetVectorMask<int16_t, MaskMode::COUNTER>(attenMaskSizeAlign / 2);
